+++ template_annotation/CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.6.0
 - Added `Method`, `Getter` and `Setter` annotations.
+- Added type specification support on `Template` annotation.
 
 # 1.5.0
 - Removed `Member` annotation.